requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
edge-tts==6.1.10
requests==2.31.0
Pillow==10.4.0
moviepy==1.0.3
numpy==2.2.5
python-dotenv==1.0.1
Jinja2==3.1.4
pydantic==2.7.1
pydub==0.25.1
aiofiles==23.2.1
httpx==0.27.0
python-multipart==0.0.9
imageio==2.36.1
imageio-ffmpeg==0.5.1
decorator==4.4.2
proglog==0.1.10
tqdm==4.66.4
